requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
PyMySQL==1.1.1
pytesseract==0.3.10
Pillow==10.3.0
python-dotenv==1.0.1
Flask-Login==0.6.3
Werkzeug==3.0.1
